refactor: Move RemountProgress component into its own file
